Module forge_fmt::helpers

Structs

Parsed

Result of parsing the source code

Functions

format

Parse and format a string with default settings

format_diagnostics_report

Formats parser diagnostics

format_to

Format parsed code

import_path_string

offset_to_line_column

Converts the start offset of a Loc to (line, col)

parse

Parse source code.

parse2

Parse source code with a path for diagnostics.

parse_raw

Parse source code, returning a list of diagnostics on failure.