Module forge_fmt::solang_ext

Extension traits and modules to the [solang_parser] crate.

Modules

ast_eq 🔒

loc 🔒

pt

Same as [solang_parser::pt], but with the patched CodeLocation.

safe_unwrap 🔒

Traits

AstEq

Check if two ParseTrees are equal ignoring location information or ordering if ordering does not matter

CodeLocationExt

Returns the code location.

SafeUnwrap

Trait implemented to unwrap optional parse tree items initially introduced in hyperledger/solang#1068.