forge_fmt::solang_ext

Module ast_eq

Macros

ast_eq_field 🔒

derive_ast_eq 🔒

gen_ast_eq_enum 🔒

wrap_in_box 🔒

Traits

AstEq

Check if two ParseTrees are equal ignoring location information or ordering if ordering does not matter

Functions

filter_params 🔒

Helper to filter [ParameterList] to omit empty parameters

to_num 🔒

Helper to convert a string number into a comparable one

to_num_reversed 🔒

Helper to convert the fractional part of a number into a comparable one. This will reverse the number so that 0’s can be ignored