forge_fmt::state

Module sol

Macros

get_span 🔒

Structs

AttributeCommentMapper 🔒

Helper struct to map attributes to their associated comments in function headers.

AttributeInfo 🔒

Decision 🔒

Enums

AttributeKind 🔒

BinOpGroup 🔒

MemberOrCallArgs 🔒

Traits

BinOpExt 🔒

Functions

get_callee_head_size 🔒

Calculates the size the callee’s “head,” excluding its arguments.

has_complex_successor 🔒

is_binary_expr 🔒

is_call 🔒

is_call_chain 🔒

is_call_with_opts_and_args 🔒

item_needs_iso 🔒

Returns true if the item needs an isolated line break.

stmt_needs_semi 🔒

Type Aliases

AttributeCommentMap 🔒