forge_fmt

Module string

Helpers for dealing with quoted strings

Structs

QuoteStateCharIndices

An iterator over characters and indices in a string slice with information about quoted string states

QuotedRanges

An iterator over the indices of quoted string locations

Enums

QuoteState

The state of a character in a string with quotable components This is a simplified version of the actual parser as we don’t care about hex or other character meanings

Traits

QuotedStringExt

Helpers for iterating over quoted strings