forge_fmt

Module comments

Structs

CommentStateCharIndices

An Iterator over characters and indices in a string slice with information about the state of comments

CommentWithMetadata

Comment with additional metadata

Comments

A list of comments

NonCommentChars

An Iterator over characters in a string slice which are not a part of comments

Enums

CommentPosition

The comment position

CommentState

The state of a character in a string with possible comments

CommentType

The type of a Comment

Traits

CommentStringExt

Helpers for iterating over comment containing strings