Trait forge_fmt::visit::Visitable

pub trait Visitable {
    // Required method
    fn visit<V>(&mut self, v: &mut V) -> Result<(), V::Error>
       where V: Visitor;
}

All [solang_parser::pt] types, such as [Statement], should implement the Visitable trait that accepts a trait Visitor implementation, which has various callback handles for Solidity Parse Tree nodes.

We want to take a &mut self to be able to implement some advanced features in the future such as modifying the Parse Tree before formatting it.

Required Methods

fn visit<V>(&mut self, v: &mut V) -> Result<(), V::Error>
where V: Visitor,

Object Safety

This trait is not object safe.